wip: vecadd steel thread w/ debuggable C++
Runs the `vecadd` example, and drastically improves the output presented
when something in WASM-land aborts execution: instead of failing
silently, we now print an error message.

Previously, that would only be shown in the dev console, and the
location of the abort would be extremely unclear and hard to
interrogate. With this commit, as long as you've got exactly the same
layout of git directories, ignored tmp symlinks, and uncommitted
build environments that I do for Talvos, the Chrome devtools work
effectively for symbolicating and debugging such conditions.