feat: step debugger
This change integrates the talvos debugging functionality into the
browser environment; it's a fairly significant change on the
Talvos side as well to support the new async-driven `Session` flow (as
opposed to the blocking-I/O-driven `stdin` prompt-style flow that
existed previously).

Unfortunately, that project is still a bit of a mess, so you'll just
have to decompile it from the `.dwp` if you want to see more details for
now.