feat: start splitting up tcf file (w/ ENTRY)
The TCF (talvos command file) is doing three things for us:

1. Defining inputs and outputs; in the case of vecadd, that's vectors a,
   b, & c. Besides allocating storage for these, it's also populating
   them (SERIES / FILL) and presenting them (DUMP).
2. Invoking the kernel with a particular "shape": MODULE points to the
   kernel sources (which is superceded by our editor), ENTRY says where
   to jump into that module, and DISPATCH invokes it.
3. Binding the first two together; that's what (I believe)
   DESCRIPTOR_SET is doing, and that's what would happen with multiple
   ENTRYies/DISPATCHes referencing the same set of descriptors/buffers.

(It's also got a fourth job in talvos-land: the `# CHECK` comments serve
as a "blessed" output that the testing framework looks for, but we're
not doing anything with those.)

The first two, and especially how they map to each other, are key
intutions for learners to build, so we've got to take over those
behaviors to try and find a more accessible way to build that
understanding. The last one I don't understand well enough yet, but I
suspect it's accidental complexity from Talvos' particular construction
that we ought to be able to roll into the other two.

Either way, this change starts the process of unwinding the TCF spring
and mapping it into what are ultimately DOM elements in the browser. The
particular mechanism for doing this is an experimental approach to try
and solve these two related challenges:

1. (transparency) Exposing details of Talvos' model through the
   WASM<->JS barrier right now is a lot of rote steps; the tools that I
   was able to evaluate (embind, WebIDL) for smoothing this part out
   didn't seem to offer very much utility (mostly requiring the same
   rote steps, just in *their* bespoke DSL). Instead, this approach
   attempts to read the internal model state as directly as possible, so
   that minimal source changes (/ spurious copies) are required to bind
   to a new property.

2. (efficiency) Briefly; the WASM "linear memory" is embedded in the
   javascript heap. The browser's execution engine, therefore, can read
   & write it "for free," but anything that we want Talvos to operate
   over needs to be inside that segment for it to be visible. So, we
   *simply* choose to allocate shared things inside the linear memory
   directly, and pass around references to avoid copies!

Curious readers might note that our strategy implies that now JS-land
and C++-land have to agree about the representation of various data
structures & objects; a challenge that neither ecosystem accepts
willingly. But, ABI stability is just a social contstruct, after all.
Given the very specific build conditions we're using on the talvos side
(a pinned libc++, clang, emscripten, the works), we should at least have
warning when it's being broken, and options around timing of uptaking
changes.

Conditions under which `lib/binding` might be removed include:

- A better answer from the broader WebAssembly community; this is an
  area that's actively being worked on, but it's a little impenetrable
  to me at the moment whether any of the answers meet our needs and how to
  use them.
- Re-writing Talvos in javascript to eliminate WebAssembly as a
  dependency; that would be a shame as this is (or, ought to be)
  WebAssembly's sweet spot.
- If the maintenance burden ends up being too high, then re-writing the
  mapping in either aforementioned tool (embind, WebIDL) ought to also
  do the job.

Also included in this too-big change are:

1. An experiment with memory benchmarking that ~kinda works,
2. Efforts to understand what kinds of JS v8 finds easier or harder to
   execute; unfortunately, the resolution of the `tinybench` framework
   that's built-in to `vitest` is only 1ms, so it's waaaayyyyy too
   coarse to be super useful to us.
3. A wild hair about implementing a teensy-tinsy (~1 instruction)
   function in WASM to speed up bit-twiddling in `BitSet`; sorry, when I
   get sad I make thinks faster and this was a sad week for me.