build: hoist entire wasm build into this project
Including both talvos and its dependencies. Now, we ought to be able to
produce the entire tool from a single (submodule-recursive) clone of
this project alone, without requiring external setup/pre-build steps.

Also, prior to this snapshot, invoking the build-wasm script in watch
mode wouldn't watch for changes when the initial first-pass build
failed. Now, it will watch a failing build, at the cost of moving the
main flow "out of line."