feat(vis, wip): step buttons for the new wip vis
The core idea here is to identify the specific quanta we're trying to
model. In order to really communicate what is simultaneous about SIMT
(and how it differs from something like SIMD, which expresses a distinct
_kind_ of simultenaity), it seems we must split out "starting" an
operation from "completing" it, so this draws out a visual model for
that to experiment with before we go into talvos and make it real.