feat(wip): add rough cut at result text to vis
The trouble is that most results are going to be both variable-width and
likely to require multiple context-dependent visualizations.

The main benefit of seeing the upcoming results, i.e. before they have a
chance to be computed, and therefore before we can know their width(s),
is to help the person "wayfind" in program-space, and so they get a sense
for what's coming up next.

We might:

1. Move the computed results to a different part of the view, so the
   stream of operations is its own linear context,
2. Add an "on-deck" or "minimap" kind of view to help localize without
   necessarily needing to draw every upcoming operation,
3. Re-flow the layout after we've computed a particular result (in
   addition being hard with SVGs, that also has the side effect of
   making the view very "jumpy"),
4. Define a maximum size (good) and always size our view to that (bad)