feat: new "dot-matrix" view of the history
Improves signalling around in-flight operations, and leaves behind a
record of parallelism. Attempts to afford a linkage between the source
lines and the result "blocks" by using similar iconography, and
including a caret in the "gutter" (all just ascii art, for now).

Also includes a few fixes (sorry) for usage patterns like
tick-tick-step; "all ticks 'til the end"; and ticking/stepping off the
end of the program.